pyproject.toml
[project]
name = "ai-agents-workflows"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
